merry xmas

+++ srcs/ft_putnbr.c
@@ -6,15 +6,15 @@
 /*   By: ssabbah <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/19 13:52:02 by ssabbah           #+#    #+#             */
-/*   Updated: 2017/12/13 15:40:35 by ssabbah          ###   ########.fr       */
+/*   Updated: 2017/12/24 18:01:32 by ssabbah          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(long long n)
 {
-	long int nb;
+	long long nb;
 
 	nb = n;
 	if (nb < 0)
